refactor: Moving custom classes tests to server side
